feat(helper): add isSameGroupPosType
