+++ packages/textlint-report-helper-for-google-preset/src/textlint-report-helper-for-google-preset.ts
@@ -1,95 +1,11 @@
 // MIT © 2017 azu
 import { matchTestReplace, TestMatchReplaceReturnDict } from "match-test-replace";
-import { EnglishParser } from "nlcst-parse-english";
-import { Tag } from "en-pos";
 import { ASTNodeTypes } from "@textlint/ast-node-types";
 
 const { RuleHelper, IgnoreNodeManager } = require("textlint-rule-helper");
-const toString = require("nlcst-to-string");
 const StringSource = require("textlint-util-to-string");
-const findUnistNode = require("unist-util-find");
-const parser = new EnglishParser();
 
-// Additional lexicon
-const lexicon = require("en-lexicon");
-lexicon.extend({
-    browser: "NN"
-});
-
-/**
- * Pos Type
- * @see https://github.com/finnlp/en-pos
- */
-export const PosType = {
-    Noun: "NN",
-    PluralNoun: "NNS",
-    ProperNoun: "NNP",
-    PluralProperNoun: "NNPS",
-    BaseFormVerb: "VB",
-    PresentFormVerb: "VBP",
-    PresentForm3RdPerson: "VBZ",
-    GerundFormVerb: "VBG",
-    PastTenseVerb: "VBD",
-    PastParticipleVerb: "VBN",
-    ModalVerb: "MD",
-    Adjective: "JJ",
-    ComparativeAdjective: "JJR",
-    SuperlativeAdjective: "JJS",
-    Adverb: "RB",
-    ComparativeAdverb: "RBR",
-    SuperlativeAdverb: "RBS",
-    Determiner: "DT",
-    Predeterminer: "PDT",
-    PersonalPronoun: "PRP",
-    PossessivePronoun: "PRP$",
-    PossessiveEnding: "POS",
-    Preposition: "IN",
-    Particle: "PR",
-    To: "TO",
-    WhDeterminer: "WDT",
-    WhPronoun: "WP",
-    WhPossessive: "WP$",
-    WhAdverb: "WRB",
-    ExpletiveThere: "EX",
-    CoordinatingConjugation: "CC",
-    CardinalNumbers: "CD",
-    ListItemMarker: "LS",
-    Interjection: "UH",
-    ForeignWords: "FW",
-    Comma: ",",
-    MidSentPunct: ":",
-    SentFinalPunct: ".",
-    LeftParenthesis: "(",
-    RightParenthesis: ")",
-    PoundSign: "#",
-    CurrencySymbols: "$",
-    OtherSymbols: "SYM",
-    EmojisEmoticons: "EM"
-};
-
-export const getPosFromSingleWord = (word: string): string => {
-    const tags = new Tag([word])
-        .initial() // initial dictionary and pattern based tagging
-        .smooth().tags; // further context based smoothing
-    return tags[0];
-};
-
-export const getPos = (text: string, word: string): string => {
-    const CST = parser.parse(text);
-    const node = findUnistNode(CST, (node: any) => {
-        if (node.type === "WordNode") {
-            return toString(node) === word;
-        }
-        return false;
-    });
-    if (!node) {
-        return "";
-    } else if (node.data && node.data.pos) {
-        return node.data.pos;
-    }
-    return "";
-};
-
+export { getPos, getPosFromSingleWord, PosType, isSameGroupPosType } from "./en-pos-util";
 // str
 export const shouldIgnoreNodeOfStrNode = (node: any, context: any) => {
     const helper = new RuleHelper(context);
