+++ packages/textlint-report-helper-for-google-preset/test/en-pos-util-test.ts
@@ -1,0 +1,1 @@
+// MIT © 2017 azu
